// src/types/api.ts
export interface ImageResponse {
  id: string;
  imageUrl: string;
  publicId?: string;
}
// ...
export interface BannerRequest {
  title: string;
  subtitle?: string;
  description?: string;
  buttonText?: string;
  buttonUrl?: string;
  displayOrder?: number;
  active?: boolean;
}

export interface BannerResponse {
  id: string;
  title: string;
  subtitle: string;
  description: string;
  buttonText: string;
  buttonUrl: string;
  displayOrder: number;
  active: boolean;
  image: ImageResponse | null;
  createdAt: string;
  updatedAt: string;
}